bug 1218576 - Ensure accumulations are for valid IDs. r=gfritzsche
We're seeing occasional failures of the isValid assert on the parent side where
the ID we're being asked to accumulate on is invalid.

Check the IDs on both sides to ensure that they're valid heading in.

MozReview-Commit-ID: 4gLG54Tx64s

UltraBlame original commit: 1b498933da3ae553f8a965c4b20e40fbe57fd1bf