Bug 1304176 - Document and refactor query_virtualenv_path; r=ted
We don't need a variable to hold the result. Just use return.

The "virtualenv_path" option has a default value, so it should always
be set. Add code confirming that. And refactor the code to use
less indentation. And remove a branch that can never occur since
the virtualenv path is guaranteed to be defined.

MozReview-Commit-ID: DZ6LnlxZJFj

UltraBlame original commit: d5355738ce1edf58cabee849024c7bd02c641b77