Bug 1304302 part 6 - Add cycle collecting support to pointer of StyleSheet. r=smaug
This is mostly copied from StyleSheetHandle which is being removed in
later patches. This is a temporary workaround to keep things working.
It should be removed soon when StyleSheet itself becomes a subclass
of nsISupports (which is expected to happen in bug 1292432).

MozReview-Commit-ID: Gqv0eRZgja5

UltraBlame original commit: 203b90fcacdf0092fc0867b7afeaf3c424745eae