Bug 1286900 - Ensure WPT tests have a version control checkout; r=ted
We add a mozharness action to the TestingMixin base class that ensures
we're running from a VCS checkout and we add this action to the WPT
script.

This ensures that we always have a source checkout available, even
in buildbot. (Before, we only had a source checkout in TaskCluster.)

MozReview-Commit-ID: 26NxwDZywXr

UltraBlame original commit: 01c3527edde5d90114b72e52e5890b43e6ba15cb