Bug 1306421 - Don't use proxxy if running in TaskCluster; r=armenzg
Proxxy is only configured in buildbot land. Don't enable it in
TaskCluster.

Ideally, we'd only enable proxxy if we detect we're in a buildbot
environment. But the change in this commit is more conservative
and aligns with existing behavior.

MozReview-Commit-ID: HBCdQ6MkYGa

UltraBlame original commit: 7de011ec6e45a2e5019ecae87acacb8bb2a8f381