bug 1306662 - Make desktop-test and desktop1604-test pull minidump_stackwalk binary from the in-tree tooltool manifest. r=gps
Instead of having the desktop[1604]-test images currently bake in a
minidump_stackwalk binary from some random S3 URL, use the binary listed
in the in-tree tooltool manifest that we use for buildbot test jobs. As
a nice side-effect, this will ensure that the desktop-test images get
rebuilt whenever that manifest is updated with a new version, so they
will continue to use the right version in the future.

MozReview-Commit-ID: 6bThddwq6p1

UltraBlame original commit: 22eee779473b155f28aab3fd9a5248d1d0646a56