Bug 1307009 - Restrict the CSS properties db test to only nightly; r=tromey
The CSS properties DB breaks on uplift, and the only way to extract the
information is through a full build. Since the DB is used only as a
back-up if the CssPropertiesActor does not provide full information,
it's ok to include experimental properties even in release.

MozReview-Commit-ID: 7aDIsyZjf6h

UltraBlame original commit: 04b5cc7a68efbed9fb044c4ab45f90d6db47dc26