Bug 1223658 - Part 4: Add a function to check active duration is zero. r=birtles,dholbert
We are going to pass animations to the compositor in before phase but not to
pass when active duration is zero.  To distinguish this state we need a new
function to check that the active duration is zero.

MozReview-Commit-ID: 2Lckmt4LQcV

UltraBlame original commit: a523efc645d048fec7f68c32c4c3975e882e8160