Bug 1292071 - Verify VCS checkout is pristine; r=dustin
The working directory state in automation is somehow incorrect. To
help debug what is doing on, we increase the verbosity of the
`hg robustcheckout` command and perform an explicit `hg status`
after `hg robustcheckout` and verify that the working directory has
no differences from the pristine checkout state.

`hg status` can be expensive. So this is meant as a temporary
stop-gap until we can figure out the underlying problem.

MozReview-Commit-ID: LIHmoMlxNsG

UltraBlame original commit: 4d50a677b98973910b94eddf4753c2209d017efc