Bug 1310509 - Call NotifyAsyncPanZoomStarted/Stopped on document containing current scrolling content. r=kats
We now call NotifyAsyncPanZoomStarted/Stopped precisely on the document
which is being transformed, so we no longer need to notify the child
docshells which was added in Bug 1088559.

Remove the |nsIDocument| argument for ProcessAPZStateChange(), which is not
used anymore.

mActiveAPZTransforms added in bug 1142926 is removed because AccessibleCaret
is the only consumer for AsyncPanZoomStarted/Stopped, and it now defaults to
always show while scrolling, i.e.
"layout.accessiblecaret.always_show_when_scrolling" defaults to true. And I
cannot reproduce the bug even if I turn off the preference.

MozReview-Commit-ID: DiEk2gCIHn2

UltraBlame original commit: 6fbbd157345fc54d39e3265865c45963df5243fa