Bug 1313477 - Convert XPCOM test TestObserverArray to a gtest. r=froydnj
This is a very straightforward conversion of TestObserverArray. The tests were
not split out as they are all dependent on each other. The test macro was left
as well as converting it to a proper function would require a fair amount of
modification to the tests.

MozReview-Commit-ID: Bz9aWbAwBS

UltraBlame original commit: b64e935b6316c2ebae0566d0577f0996d9729fc5