Bug 418833 - Bump fuzzyness on ua-style-sheet-checkbox-radio-1 reftest for Android. r=jrmuizel
I talked to mstange about this, and what might be happening here is that there's
a difference in rounding going on during (I think) rasterization. The change is
very small and not human-noticable, so I think taking this fuzzyness is worth
the cost considering the gain in functionality.

MozReview-Commit-ID: C0CPNrIdCDu

UltraBlame original commit: 5df1453ded4cb1603ca35ae7a36330108ead5cfa