Bug 797208 - Allow the stub installer to install either 32-bit or 64-bit builds, and make 64-bit the default when supported; r=rstrong
A new control allows the user to select 32 or 64-bit when the system supports both,
and it defaults to 64-bit when available. This means the stub installer is now
the same regardless of its build architecture; it was always a 32-bit executable
anyway, but now its actual behavior depends only on the running system, not the
target architecture of the application it was built alongside.

The options screen has been rearranged according to a design by Michael Verdi
so that the new control doesn't leave the UI so badly cluttered.

Also removed TmpVal, which wasn't used in the stub and so was generating warnings.

MozReview-Commit-ID: 5baJCkAa7bJ

UltraBlame original commit: 8638691b8a7a641b6d40eb4f59c95920c38b188a