Bug 1320399 - Fix error in Grip Rep; r=nchevobbe
Check that the `preview` property is not null before using it.
The bug was harmless because we catch it and return an
empty array as the properties of the grip, which is the same
when we don't have anything in the preview prop.

MozReview-Commit-ID: COU8XLATw7x

UltraBlame original commit: d65261aaa7cd0084288f1135389dfb83f9ba5058