Bug 1318697 - Part 1: Skip composing styles for properties depending on cascade level. r=birtles
When we are composing style for the Animations level of the cascade,
if we have transitions-level animations, we *also* need to compose them if we
have animations-level animations that build on top of them using additive or
accumulative composite modes.

However, we should not build those transitions-level animations unless they will
be built on or overridden by a regular animations-level animation. Otherwise
we will end up with transitions-level animations in the animations-level and
while transitions-level will override the animations-level in the cascade,
once the transition finishes there will be nothing to remove the cached
animations-level animation rule.

MozReview-Commit-ID: LaRabzDSsO5

UltraBlame original commit: 940a214b71f0b61c48309f70f90b675e34f33860