Bug 1317714 - port mediaconduit_unittests to xul gtest; r=jesup, glandium
The test results were updated to match current behaviour. The
TestDummyAudioWithTransport and TestDummyVideoWithTransports are disabled due
to shutdown crashes and intermittent failures that show up in automation.
A follow up bug has been filed to fix these. The GMP test was removed
completely as it seems unlikely that it will be practical to test that from a
gtest.

MozReview-Commit-ID: 2pOb7u2Qp7v

UltraBlame original commit: b5301d5a9251625788a510333fea870ac4a26b5b