Bug 1283898 - Put rust in path for Spidermonkey automation. r=gps
When rust is enabled, the Spidermonkey package tests need to
be able to find the rust toolchain. It's available, because
the job uses the linux64/releng.manifest, but the corresponding
path directives from mozconfig are not imported. So we set
them manually here, falling back to the rustc and cargo in
PATH, like we do for the C++ toolchain.

MozReview-Commit-ID: 8lZSQ0Ffr0c

UltraBlame original commit: 32b4d0c6aef9d711946701dcbb0b83cc0ddb5b11