Bug 1291366 - Part 1: Use GENERATED_FILES to produce AppConstants.java. r=gps
This patch lays the groundwork for two things.  First, it paves the
way for splitting AppConstants.java into two parts, a GeckoView part
and a Fennec part.  This is necessary because the Makefile.in
preprocessing is not flexible enough to write two separate GeckoView
and Fennec constants files into different directories.

Second, this allows us to more flexibly generate the file contents.
Gradle has a way to get compile-time constants into Java code, which
we want to migrate to.  The details don't matter right here, but this
paves the way to move from preprocessing to generating the
Gradle-style BuildConfig files while we continue to support both build
systems.

MozReview-Commit-ID: 2o8X99uLoaM

UltraBlame original commit: 420ee7d102653d400610a2c5add6721c100b9123